pyproject.toml
[project]
name = "in-context-learng-robotics-demo"
version = "0.1.0"
description = "InCoRo: In-Context Learning for Robotics Control Demo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.30.0",
    "opencv-python>=4.13.0.92",
    "python-dotenv>=1.2.2",
]
